+++ config/typedoc.base.mjs
@@ -1,0 +1,11 @@
+/** @type {import('typedoc').TypeDocOptions} */
+export default {
+  entryPoints: ["../src/index.ts"],
+  readme: "none",
+  includeVersion: true,
+  excludeExternals: true,
+  excludeInternal: false,
+  jsDocCompatibility: false,
+  excludeTags: ["@knipignore"],
+  sort: ["alphabetical", "kind"],
+};