+++ packages/alfa-act/src/expectation/interview.ts
@@ -1,4 +1,3 @@
-import { Future } from "@siteimprove/alfa-future";
 import type { Hashable } from "@siteimprove/alfa-hash";
 import { Option } from "@siteimprove/alfa-option";
 
@@ -24,6 +23,16 @@ type Depths = [-1, 0, 1, 2];
  * The QUESTION type maps questions' URI to the expected type of answer, both as
  * a JavaScript manipulable representation (T), and an actual type (A).
  * The SUBJECT and CONTEXT types are the subject and context of the question.
+ *
+ * @remarks
+ * That is, an Interview is either:
+ * * an ANSWER.
+ * * A Question, expecting an Interview, provided its URI, answer type
+ *   (QUESTION[URI][1]), and answer type representation (QUESTION[URI][0]) are correct.
+ *   The returned interview has depth one smaller.
+ *
+ * The complex object keys mapping ensure that the question's URI exist in the
+ * metadata.
  *
  * @public
  */
@@ -67,8 +76,8 @@ export namespace Interview {
    *
    *  Oracles must return Options, to have the possibility to not answer a given
    *  question (by returning None).
-   *  Oracles must return Futures, because the full interview process is
-   *  essentially  async (e.g., asking through a CLI).
+   *  Oracles must return Promises, because the full interview process is
+   *  essentially async (e.g., asking through a CLI).
    *
    *  The final result of the interview is either a conclusive finding with an
    *  answer, or  an inconclusive finding with a diagnostic explaining why an
@@ -93,14 +102,14 @@ export namespace Interview {
     rule: Rule<INPUT, TARGET, QUESTION, SUBJECT>,
     oracle: Oracle<INPUT, TARGET, QUESTION, SUBJECT>,
     oracleUsed: boolean = false,
-  ): Future<Finding<ANSWER>> {
+  ): Promise<Finding<ANSWER>> {
     if (interview instanceof Question) {
-      let answer: Future<Option<Interview<QUESTION, SUBJECT, TARGET, ANSWER>>>;
+      let answer: Promise<Option<Interview<QUESTION, SUBJECT, TARGET, ANSWER>>>;
 
       if (interview.isRhetorical()) {
-        answer = Future.now(Option.of(interview.answer()));
+        answer = Promise.resolve(Option.of(interview.answer()));
       } else {
-        answer = oracle(rule, interview).map((option) =>
+        answer = oracle(rule, interview).then((option) =>
           option
             // Record that the oracle was successfully used
             .tee((_) => (oracleUsed = true))
@@ -111,19 +120,21 @@ export namespace Interview {
         );
       }
 
-      return answer.flatMap((answer) =>
+      return answer.then((answer) =>
         answer
           // Recursively conduct an interview
           .map((answer) => conduct(answer, rule, oracle, oracleUsed))
           // If we still don't have a final answer, the finding is inconclusive,
           // return the last diagnostic.
           .getOrElse(() =>
-            Future.now(Finding.inconclusive(interview.diagnostic, oracleUsed)),
+            Promise.resolve(
+              Finding.inconclusive(interview.diagnostic, oracleUsed),
+            ),
           ),
       );
     }
 
     // The interview is not a question, so it is a conclusive finding.
-    return Future.now(Finding.conclusive(interview, oracleUsed));
+    return Promise.resolve(Finding.conclusive(interview, oracleUsed));
   }
 }