+++ packages/alfa-callback/test/callback.spec-d.ts
@@ -4,24 +4,17 @@ import { Callback } from "../src/callback.ts";
 
 describe("Callback", () => {
   it("defaults to void function with no extra args", () => {
-    type MyCallback = Callback<number>;
-    const fn: MyCallback = (x) => {
-      console.log(x); // This won't actually be called since this is a type test.
-    };
-    expectTypeOf(fn).toEqualTypeOf<(value: number) => void>();
+    expectTypeOf<Callback<number>>().toEqualTypeOf<(value: number) => void>();
   });
 
   it("can change return type with second generic", () => {
-    type MyCallback = Callback<number, string>;
-    const fn: MyCallback = (x) => x.toString();
-    expectTypeOf(fn).toEqualTypeOf<(value: number) => string>();
+    expectTypeOf<Callback<number, string>>().toEqualTypeOf<
+      (value: number) => string
+    >();
   });
 
   it("can accept extra args with third generic", () => {
-    type MyCallback = Callback<number, string, [boolean, Date]>;
-    const fn: MyCallback = (x, flag, date) =>
-      flag ? date.toISOString() : x.toString();
-    expectTypeOf(fn).toEqualTypeOf<
+    expectTypeOf<Callback<number, string, [boolean, Date]>>().toEqualTypeOf<
       (value: number, flag: boolean, date: Date) => string
     >();
   });