+++ packages/alfa-toolchain/src/validation/validate-all.ts
@@ -7,7 +7,10 @@ import { getPackages } from "@manypkg/get-packages";
 import * as path from "node:path";
 import { loadJSON } from "../common.ts";
 
-import { hasExtractorConfig } from "./has-extractor-config.ts";
+import {
+  hasAPIExtractorConfig,
+  hasTypedocConfig,
+} from "./has-extractor-config.ts";
 import { isInClusters } from "./is-in-clusters.ts";
 import { validateChangesets } from "./validate-changesets.ts";
 import { validatePackageJson } from "./validate-package-json.ts";
@@ -37,7 +40,13 @@ export async function validate(rootDir: string) {
 
   if (config["has-api-extractor-config"] ?? false) {
     for (const pkg of packages.packages) {
-      errors.push(...hasExtractorConfig(pkg.packageJson.name, pkg.dir));
+      errors.push(...hasAPIExtractorConfig(pkg.packageJson.name, pkg.dir));
+    }
+  }
+
+  if (config["has-typedoc-config"] ?? false) {
+    for (const pkg of packages.packages) {
+      errors.push(...hasTypedocConfig(pkg.packageJson.name, pkg.dir));
     }
   }